Move ImportQueue into own file
